+++ src/components/footer/index.tsx
@@ -1,15 +1,6 @@
 import { Contact } from "@components/contact";
 
 export const Footer: React.FC = () => (
-    // <footer className="bottom-0 mt-10 pb-10 flex w-full flex-col items-center">
-    //     <Contact />
-    // <span className="font-rubik text-center mx-10 font-light text-eerie dark:text-cultured">
-    //     Powered by the Next JS, Notion API, Framer motion and TailwindCSS.
-    // </span>
-    // <span className="font-rubik text-center mx-10 font-light text-eerie dark:text-cultured ">
-    //     Fonts courtesy of Fontshare and Google Fonts
-    // </span>
-    // </footer>
     <footer
         className="bottom-0 py-6 
         flex flex-col items-start px-4